instrument/daq: Check that self.tempdir has been set before calling os.path.isdir()
self.tempdir is initialized to None, and os.path.isdir() throws an
exception if you don't pass it a str.  This can happen if teardown()
is called before get_data(), which WA sometimes does.  Check that
self.tempdir has been initializing before calling os.path.isdir().

+++ devlib/instrument/daq.py
@@ -156,5 +156,5 @@ class DaqInstrument(Instrument):
     def teardown(self):
         self.daq_client.close()
         if not self.keep_raw:
-            if os.path.isdir(self.tempdir):
+            if self.tempdir and os.path.isdir(self.tempdir):
                 shutil.rmtree(self.tempdir)
